pyproject.toml
[tool.poetry]
name = "schalotte_tools"
version = "0.1.0"
description = "Tools for the Schalotte project"
authors = ["Tristan Weis <mail@tristanweis.com>"]
readme = "README.md"
license = "GPL-3.0-or-later"
packages = [{include = "schalotte_tools"}]

[tool.poetry.dependencies]
python = "^3.11"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
flake8 = "^7.3.0"
fake-bpy-module-latest = "^20251003"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
